lightningd/funding_tx: output number is 16 bit.
It's actually always 0 or 1, but the spec wants 16 bits, so use that here.

+++ lightningd/funding_tx.c
@@ -13,7 +13,7 @@
 #endif
 
 struct bitcoin_tx *funding_tx(const tal_t *ctx,
-			      u32 *outnum,
+			      u16 *outnum,
 			      const struct utxo **utxomap,
 			      u64 funding_satoshis,
 			      const struct pubkey *local_fundingkey,
